pyproject.toml
[project]
name = "screener-automation"
version = "0.1.0"
description = "Automated stock screener for TradingView watchlist generation"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "yfinance>=0.2.30",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "lxml>=4.9.0",
    "html5lib>=1.1",
]

[project.scripts]
screener = "run_weekly:main"
